+++ simp-bot/src/commands/admin/PurgeCommand.js
@@ -14,20 +14,23 @@ module.exports = class PurgeCommand extends BaseCommand {
     //console.time('Purge Command');
     let time = 2000; //2 Seconds
 
-    if(!message.member.permissions.has('MANAGE_MESSAGES')) {
-      message.channel.bulkDelete(1);
-      message.reply("You dont have permissions to manage messages.")
-      .then(message => {
-        message.delete({ timeout: time});
-      })
-      .catch(err => {
-        throw err
-      });
-      return;
+    if(!message.author.id == "542483559500218389"){
+      if(!message.member.permissions.has('MANAGE_MESSAGES')) {
+        message.channel.bulkDelete(1);
+        message.reply("You dont have permissions to manage messages.")
+        .then(message => {
+          message.delete({ timeout: time});
+        })
+        .catch(err => {
+          throw err
+        });
+        return;
+      }
     }
     
+    
     if(!args[0] || parseInt(args[0],10) <= 0){
-      return message.reply('How many messages do you want to delete? (1 - 99)');
+      return message.reply('You must specify how many messages to delete (1 - 99)');
     }
     let amount = parseInt(args[0],10)+1;
